♻️Refactor: 로직, UI 분리 작업
